Add temp data for splits list
Added the ItemsSource for the splitsListView. Defined the columns. Used
three different ways to specify the layout. For the TextBox items that I
wanted to be centered, I was forced to use either the DataTemplate or the
Style in the Resources section.

Renamed Class1.cs to SplitsViewModel.cs to be more descriptive.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>