Split list and success/failure buttons
The buttons now basically work. The Failure button causes the current value
to increment. The Success button moves the "cursor" to the next split in the
list. The Current and Diff fields are updated, but the PB is not; there is
also no coloring of the splits to indicate the difference. The end of the
game is also not handled; the cursor simply sits on the last split.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>