WIP moving to a simpler split edit process
I decided to start simply. The user creates a challenge using a standard
text box, with each line being a split. The View Model extracts the splits
from the text box string and creates the real list of splits.

Right now any change causes everything to reset. This isn't ideal, but it
will work for now.

Still need to handle when the user cancels the challenge creation.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>