WIP Edit splits via text box
Changed the challenge window to be a dialog, which makes more sense because
there is no need to interact with the main window at the same time.
Added OK and Cancel buttons to the dialog.
The OK click handler calls a new ViewModel method SetSplits() to give the
user text to the main program. This replaces the 'set' on the splitTextList
property.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>