Add first unit tests
Created a project for View Model unit tests. Added a few simple tests for
SuccessProc(), FailureProc(), and SetSplits().

Had to move to NotifyPropertyChanged() in order for the unit test to run.
Using the previous method, an exception was thrown, probably because there
wasn't something set up as there is when there is a View binding to it.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>