Redesign database schema
After thinking about the way the Model will be used, I decided to change the
underlying database schema. Instead of having a table for each challenge, I
now do something more standard. There are tables for Challenges, Splits, and
Runs, and they all reference each other as appropriate. This makes the database
itself a bit more complex, but hopefully it will be easier to work with.

Changed the unit test for Model. Now it just looks at the CreateChallenge
method and how it works in the common case. More unit tests will follow.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>