Replace CreateNewRun with StartNewRun
After thinking more about use cases, I decided that it made more sense for the
user to "start" a new run thant to "create" a new run. The biggest reason is
because there is no need for the user to have multiple runs in progress for
the same challenge.

Made some other minor changes:
* No need for NumberOfRuns in Challenges. This can be derived from the number
of rows in the Runs table.
* No need for RunNumber in Runs. The order is given by ID. If necessary, I
will add a DateTime.
* Added a few more explicit exception throws to make things consistent.
* Cleaned up some unneeded/deprecated code.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>