Change Model.CreateChallenge to use list of strings
Having the input parameter for CreateChallenge be a list of Split objects
didn't make sense to me. Becaue the challenge doesn't yet exist, the only
thing the caller has is a list of strings. Therefore I changed the method
to take List<string> instead of List<Split>.

This required an update to most of the unit tests. I took this opportunity
to pull the List allocation code out of each individual test method and
put it in the ClassInitialize method.

Also added a check for duplicate challenge name to CreateChallenge. I do
not know how the test passed before this.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>