WIP Change AddNewChallenge to use List<string> for splits
I finally decided I didn't like taking the "one giant string of names"
approach in the ViewModel. This only matches one specific implementation
of the View (using a free-form TextBox). Using a List<string> is more
generic and thus easier to understand. The Model has to do the conversion,
but it isn't that bad, especially since I had the code written already.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>