Implement the main database access methods of Model
The methods CreateDatabase, LoadDatabase, and Save are now fully implemented.
The Model must be given the path of the database file, as this allows for the
most flexibility. The app using the Model might, of course, decide to hard-
code the location of this file and not allow the user to choose it.

Unit tests are present and pass.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>