Implement CreateChallenge in SplitsViewModel
The CreateChallenge method is now implemented. It now takes a string for the
name and a list of strings for the split names. It updates the database and
notifies the app that the properties have changed.

The View had to change because it is now responsible for taking the splits
as they were displayed and convert them into a list of strings. The View
also now displays the name of the current challenge.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>