Implement select current challenge [issue #7]
When the user clicks on a challenge in the list of challenges, the list of
splits will be updated to show the splits for the selected challenge.

I moved the list of challenges to be next to the list of splits instead of
underneath it. This is because the list of splits resizes itself to fit the
number of items. This causes the list of challenges to change position, which
will cause another (unwanted) select event if the cursor is now on a different
item in the challenge list. This can be seen with a normal click duration, but
it can be easily reproduced by pressing and holding the mouse button on a
challenge that has a different number of splits than the current challenge.
There might be a way to fix this issue programmatically, but I decided it was
best to move the challenge list. Not only does it look better but I still think
that the challenge list will eventually be removed from the main window.

The PB value is not handled correctly yet; there is a temp value in its place.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>