Implement set current challenge when started [Issue #6]
The app will now automatically set the current challenge to whatever the user
had when the app last closed.

An Application Setting named "LastUsedChallenge" was created in the ViewModel
project. It is set and saved every time the current challenge is set. The
Setting is of scope "User" which means that the config file used to store the
value is in the user's AppData/Local folder.

I tested with a blank value (i.e. the first time the app is launched) and
after creating a new challenge. I am not able to test after deleting a
challenge, because that feature is not yet implemented.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>