Abstract out user settings; add Moq framework for testing it
SplitViewModel was directly accessing the settings database, which meant
that the unit tests and normal operation would interfere with each other.
The solution is to decouple the app from the database by abstracting the
interface to the database. The interface then is passed to the constructor,
which allows the unit test to mock it. This is apparently a common thing
to do these days; it might also need to be done for the app database file
as well, but that is a topic for another day.

The Moq framework was added to the project so that I didn't have to make
the mock completely from scratch. It looks like this is a popular framework,
and so far it has been easy to work with. The SplitViewModel unit tests have
been updated to mock the user settings as needed (they fail when run against
the previous commit but I did not have time to research a good solution).

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>