Make Homunculus_Model more modular
I am now thinking about using a mock for Model in the ViewModel unit tests.
To do this I need to make Model more modular. I looked online for some
suggestions on how to do this but I didn't find anything definitive. Thus,
I am doing it in a way that makes sense to me.

Homunculus_Model needs to publish an interface that all of its potential
Model classes will implement. This means that the interface should be in
Homunculus_Model, not some other namespace. As this would be the third thing
in Model.cs I decided to break everything out into their own file, with the
filename being the same as the definition, as seems to be the convention. As
I am looking to be more generic and future-proof, I decided that the current
model class needs a more description name, so I changed it to ModelXml. This
means that the Model.cs file needs to change names as well.

For now, the only method in IHomunculusModel is DeleteChallenge, since that
is what I am working on at the moment. Once that is complete, I will migrate
the rest of the methods to the interface. The only ones I am concerned about
are LoadDatabase and CreateDatabase. Currently they take the filename as the
only parameter. This doesn't make sense for other possible model
implementations such as a remote database, so this might need to change. It
seems that maybe this could be a generic connection string, but I want to
think about it more before I go in that direction.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>