Convert SplitsViewModel to use IHomunculusModel
The main constructor for SplitsViewModel now takes an IHomunculusModel
as a parameter. Currently, the constructor instantiates an object of
type ModelXml like it has been doing, because I still don't know how to
provide the parameters in the DataContext in the XAML.

IHomunculusModel has been updated with all of the methods that the
view model will need to use.

No unit tests have been updated to make use of this. This will come
later.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>