Implement Delete Challenge in the View
Added a delete button next to the create button.

There was an issue with the selection change on the challenge list view.
I worked around it by checking for null on SelectedItem. This feels like
a hack, but I hope it will be addressed when I fix the bug already entered
against the selected item in the list not changing when it should.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>