Add mocking of Model to ViewModel unit tests
The mock is not used by all tests at this time, so there will be more
changes to come.

With the way the Model is defined, using a mock means that the tests
need to know some of the internal details of the methods under test.
I decided that this was okay, since it means that the MUTs will not
be executing extraneous code.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>