Redesign ModifyChallenge
After thinking about it some more, I decided to divide challenge edits into
two different categories: cosmetic and structural. Method ModifyChallenge()
is for cosmetic edits only, which includes changing the name of the challenge,
changing the names of the splits, and changing the order of the splits.

Adding and deleting splits are structural changes, and they will be handled
by cloning the existing challenge and presenting the user with the same
interface as when a challenge is created. Thus structural changes require a
new challenge to be created. This avoids all of the mess with history and PBs
when some splits are not consistent.

My current plan is to do this for version 1. If I come up with a different way
to do it I will likely bump the major version number to 2 because it will no
doubt be a big difference.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>