Remove TextBox for splits in CreateChallengeWindow
The text box has been deleted. Now the splits are only shown in a ListView.
This is fine for the display, and it is even possible to change the names
of the splits. However, the ability to move a split or delete it is not
present, and the Add Split button might not be fully functional. All of this
is coming in this sprint so watch this space ->.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>