Fix end of run handling [Issue #28]
Ending a run was leaving the challenge in a bad state, such that the success
and failure buttons could subsequently be clicked and thus cause an exception
to be thrown.

Now the ViewModel notifies on every set of RunInProgress. The Model uses this
property to control when the buttons are active.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>