Fix split color change during run [Issue #32]
The split color wasn't changing correctly during a run (i.e. changing from
green to red when the PB for that split is exceeded). It turns out that a
Converter will not be triggered when the element of a list is changed.
Instead, a MultiConvert must be used, and one of the bindings must be the
list property of the ViewModel (i.e. the ItemsSource of the ListView control).

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>