Change ChallengesView to be more flexible
I was having difficulty figuring out how to communicate changes between
MainVM and ChallengesVM. I decided to try allocating a new ChallengesView
object each time a change to the global app state occurs. Currently, this
is implemented with two buttons that simulate creating a new data store and
loading an existing data store. Clicking either of these buttons causes
MainVM to change the contents of the data store, then a new ChallengesView
is created, which accesses the new data store in its constructor.

There is probably a better way to do this, but try this way for now.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>